Bugfix: Defaults should not be applied when an empty Object is specified.
When an Object is not specified, such as with "path_sources", then the default should be loaded.
When an Object is specified but ha Content with a length of 0, then the Content should not have a default applied.
The example bzip fakefile exposed this problem where the "path_sources" is empty and should be.
The existing code is incorrectly applying the default when length is 0.

The code to actually check for whether or not an Object is specified is already in place but it is not used.
Update the code to utilize the existing information.