Update: Simplify build settings files.
The build setting files contained numerous Objects set to empty content.
This allows for showing how to configure.

With a long standing bug fixed (defaults being applied incorrectly) a misconfiguration as a result of this has been revealed.

Simplify the build settings to make the settings easier to read and maintain.
This ends up fixing the bug (because the default path_sources gets assigned when path_sources is not defined at all).

Do some minor structural cleanup in these files.