Feature: Add missing functionality allowing the utf8 program to convert back to binary data with invalid codepoints.
Even when there are invalid codepoints produced, it should be possible to convert the entire output back to the original data.
This is possible because the codepoint output by default still prints the invalid data as a hex-digit representing up to 4 bytes of data.
The combining and width parameters are also supported.